+++ webapp/apps/gateway-ui/src/client/app/modules/web-client/ssh/web-client-ssh.component.ts
@@ -5,7 +5,6 @@ import { DVL_SSH_ICON, JET_SSH_URL } from '@gateway/app.constants';
 import { AnalyticService, ProtocolString } from '@gateway/shared/services/analytic.service';
 import { WebComponentReady } from '@shared/bases/base-web-client.component';
 import { TerminalWebClientBaseComponent } from '@shared/bases/terminal-web-client-base.component';
-import { ToolbarSessionInfo } from '@shared/components/floating-session-toolbar/models/session-info.model';
 import { GatewayAlertMessageService } from '@shared/components/gateway-alert-message/gateway-alert-message.service';
 import { SshConnectionParameters } from '@shared/interfaces/connection-params.interfaces';
 import { SSHFormDataInput } from '@shared/interfaces/forms.interfaces';
@@ -36,10 +35,6 @@ export class WebClientSshComponent
   @ViewChild('webSSHGuiTerminal') webGuiTerminal: ElementRef;
 
   formData: SSHFormDataInput;
-  sessionInfo: ToolbarSessionInfo = { rows: [], emptyValueText: 'N/A' };
-  private sessionInfoUrl: string | null = null;
-  private sessionInfoUsername: string | null = null;
-  private lastSessionInfoKey = '';
 
   private remoteTerminal: SSHTerminal;
   // unsubscribeTerminalEvent, unsubscribeConnectionListener, removeRemoteTerminalListener()
@@ -223,61 +218,5 @@ export class WebClientSshComponent
 
   protected getProtocol(): ProtocolString {
     return 'SSH';
-  }
-
-  private buildSessionInfo(): ToolbarSessionInfo {
-    return {
-      rows: [
-        { id: 'sessionId', label: 'Session ID', value: this.webSessionId, monospace: true, order: 1 },
-        { id: 'url', label: 'URL', value: this.sessionInfoUrl, monospace: true, order: 2 },
-        {
-          id: 'username',
-          label: 'Username',
-          value: this.sessionInfoUsername,
-          hidden: !this.sessionInfoUsername,
-          order: 3,
-        },
-      ],
-      emptyValueText: 'N/A',
-    };
-  }
-
-  private refreshSessionInfo(): void {
-    const next = this.buildSessionInfo();
-    const nextKey = this.buildSessionInfoKey(next);
-    if (nextKey === this.lastSessionInfoKey) {
-      return;
-    }
-
-    this.lastSessionInfoKey = nextKey;
-    this.sessionInfo = next;
-  }
-
-  private buildSessionInfoKey(info: ToolbarSessionInfo): string {
-    const rows = [...info.rows].sort(
-      (a, b) => (a.order ?? Number.MAX_SAFE_INTEGER) - (b.order ?? Number.MAX_SAFE_INTEGER) || a.id.localeCompare(b.id),
-    );
-
-    return JSON.stringify({
-      title: info.title ?? null,
-      emptyValueText: info.emptyValueText ?? null,
-      rows,
-    });
-  }
-
-  private toUserFacingUrl(url: string | null | undefined): string | null {
-    if (!url) {
-      return null;
-    }
-
-    try {
-      const normalized = new URL(url, window.location.href);
-      normalized.protocol = normalized.protocol === 'wss:' ? 'https:' : 'http:';
-      normalized.search = '';
-      normalized.hash = '';
-      return normalized.toString();
-    } catch {
-      return url;
-    }
   }
 }
